rollout/3: restructuring
