+++ lib/src/home/view/cards.dart
@@ -11,7 +11,7 @@ import 'package:skg_hagen/src/common/model/sizeConfig.dart';
 import 'package:skg_hagen/src/common/service/client/assetClient.dart';
 import 'package:skg_hagen/src/common/service/client/dioHttpClient.dart';
 import 'package:skg_hagen/src/common/service/network.dart';
-import 'package:skg_hagen/src/common/service/pushNotificationManager.dart';
+import 'package:skg_hagen/src/pushnotification/service/pushNotificationManager.dart';
 import 'package:skg_hagen/src/common/service/tapAction.dart';
 import 'package:skg_hagen/src/home/controller/home.dart';
 import 'package:skg_hagen/src/home/model/aid.dart';
