+++ lib/src/pushnotification/model/messaging.dart
