+++ lib/src/pushnotification/model/notification.dart
